ss_shortest_distance: fix, and speed up
As an "optimization", the todo queue of ss_shortest_distance was
doubled with a set of nodes already treated.  But this set was not
correct: the question is not whether the node has already been
treated, but if it's already scheduled for a visit, even if it has
already been visited previously.

* vcsn/algos/distance.hh: Fix style.
Use a vector instead of a map indexed by state_t.
Fix the bug.
Remove weird final assignment.
* tests/python/push-weights.py: Check the result.
* vcsn/misc/deque.hh: New.