shortest: non free labelset: new support
Several attempts were made to find something efficient.  The
free-labelset implementation was disabled, so all the following
benches are truly about the non-free-labelset implementation,
even for benches from score that normally use the free-labelset
implementation.

First attempts used one polynomial per state, instead of
pairs (monomial, state).  (1) is the free-label implementation
score, (2) is using polynomials.

  (1)    (2)
 0.77s 18.27s : a.shortest(5)        # a = de_bruijn(9), 10x
 0.83s 18.26s : a.shortest(5000)     # a = de_bruijn(9), 10x
 0.24s  1.32s : a.shortest(5)        # a = std([a-e]?{600}), c = [a-e] -> Z

Then we moved to using a heap of monomials.  I used Boost.Heap,
and in some cases it does not compile cleanly: Priority queue and
D-ary Heap.  Benches seem to prefer Binomial heaps (but in some
cases Pairing is better).

Fibonacci Binomial Skew  Pairing
 3.14s     2.77s   3.81s  2.11s  a.shortest(5)        # a = de_bruijn(9), 10x
 4.39s     3.55s   4.73s  2.65s  a.shortest(5000)     # a = de_bruijn(9), 10x
 0.99s     0.84s   3.51s  3.30s  a.shortest(5)        # a = std([a-e]?{600}), c = [a-e] -> Z

Then, instead of using monomials, lets enque (state, label, weight)

 2.77s 2.26s : a.shortest(5)        # a = de_bruijn(9), 10x
 3.55s 2.88s : a.shortest(5000)     # a = de_bruijn(9), 10x
 0.84s 0.77s : a.shortest(5)        # a = std([a-e]?{600}), c = [a-e] -> Z

But then the result was disastrous on a new bench case in score:

  a.shortest(5000) # a = derived_term((\e|x + a|\e)*), c = [a]?x[x]? -> Q, 10x

which generates an immense number of identical triples (state,
label, weight) (which is easy to foresee as anyway there is a
single state here!).  So instead of having a bazillion of such
triples, we pop from the queue all the triples that have the same
state and label, and sum their weights.  This is costly for the
first three usual benches, but it makes an immense difference for
the new one (from "I can't wait that long, what's going on?" to
0.29s).

 2.26s  2.43s : a.shortest(5)        # a = de_bruijn(9), 10x
 2.88s  3.14s : a.shortest(5000)     # a = de_bruijn(9), 10x
 0.77s  0.79s : a.shortest(5)        # a = std([a-e]?{600}), c = [a-e] -> Z
 N/A    0.29s : a.shortest(5000)     # a = derived_term((\e|x + a|\e)*), c = [a]?x[x]? -> Q, 10x

* vcsn/algos/shortest.hh: here.
* tests/python/shortest.py: Check.
* bin/vcsn-score: Bench.