automata: beware of special characters
We are still fragile on I/O of automata when the labels are a bit
special (say `'`, or `|`, or `<`, etc.).

We need to support more escapes in input, and escape more in output.

* lib/vcsn/algos/edit-automaton.cc
* vcsn/alphabets/char.hh (get_char): When a symbol is backslashed, always
strip the backslash.
(get_letter): Support quoted letters.
(print): Escape special characters.
* vcsn/alphabets/setalpha.hh (get_word): Support quoted words.
* vcsn/labelset/nullableset.hh (conv): Be ready to pass escaped character
to the inner labelset.

And we also need to be able to not escape, in the case of OpenFST.  To
this end, format::raw is introduced.  But it might be a mistake: maybe
the way things are printed should be dealt with in OpenFST itself.
Maybe the future will tell which approach is better.
* vcsn/misc/format.hh (raw): New format.

* tests/python/efsm.dir/ascii-to-one.efsm, tests/python/efsm.dir/ascii-to-one.gv:
New.
* tests/python/efsm.py: Use them to check such I/O for transducers.