efsm: prefer lan<string> to law<char>
The semantics of "words" in OpenFST is rather that of letters that are
strings.  This is what lal<string> models.

* vcsn/alphabets/string.hh: Improve conformance in input/output.
* vcsn/labelset/letterset.hh (print): Honor the format.

* vcsn/algos/edit-automaton.hh, lib/vcsn/algos/edit-automaton.cc:
Provide a means to choose a different context than the guessed one.

* lib/vcsn/algos/efsm.cc: When law<char> is guessed, prefer lan<string>.

* tests/python/efsm.dir/ascii-to-one.efsm,
* tests/python/efsm.dir/ascii-to-one.gv: Adjust.
At least for the time being, drop the case of the '<space>' token.
Indeed spaces are ignored in FSM files, so one _has_ to use a different
symbol.  Unfortunately, this makes transducers that take characters
as input look like transducers taking strings as input, _only_ because
`<space>` is used to denote ` `.
So drop it completely, until we find a cleaner means to support it.

* tests/python/efsm.dir/str.efsm, tests/python/efsm.py: Don't
try to pass law to OpenFST.
Maybe we should letterize first?