score: remove lightest
We found that lightest was too naive, and failed to find the really
smallest weights in some cases.  And besides, because an optimization
was removed in the final version of the first implementation, the
score actually takes for ever to complete, and completely eats up all
the memory of the machine it's running on.

So, until this is fixed, let's...

* libexec/vcsn-score (shortest): remove.