notebook: make it work when installed
For some reason, when `migrated` does not exist, running ipython
notebook with our profile triggers the creation of this file.  If the
installation was run by root, of course, Jupyter fails to create this
`migrated` file, and the whole process fails.  If migrated is
installed, it works.

Let's do that until we know better.

* share/vcsn/jupyter_profile/migrated: New.
* share/vcsn/local.mk: Install it.
Fix the installation directory for Jupyter/IPython files.