project: split project-automaton
We have very ugly mutual dependency issues between focus, copy, and
project.  Actually, it's project_automaton which is a problem, so
let's move it into its own header to fix this mess.

The proper fix is to move this into the appropriate headers.
Currently, core/rat uses make_proper which is in algos/.  This is
insane, we _must_ restore proper dependencies, in layers.

* vcsn/algos/project-automaton.hh: New.
* vcsn/algos/project.hh: Extracted from here.