shared_ptr: keep copies, not const-refs
Passing shared-ptrs as const-refs is fine, and actually preferrable:
we avoid useless (and noticeably coslty) inc/dec on the reference
counter.  Storing them as such is dubious and possibly harmful.

Dubious, because we keep a reference to a temporary object: the
shared-ptr (_not_ the object itself).  Harmful, because in some
situations that shared-ptr may actually be destroyed: consider for
instance lazy operations such as conjunction or composition: the
shared-ptr may have been destroyed long ago (on lazy funtion return)
when we actually need to use it (evaluation of a lazy state).  This
was noticed by Valentin Tolmer when working on lazy composition (see
ece3a63276a96b653576bf68c00c93c5e7828edc).

It would probably be nice to be able to force some of these shared-ptr
to point to const.

* vcsn/algos/are-isomorphic.hh, vcsn/algos/copy.hh,
* vcsn/algos/distance.hh, vcsn/algos/epsilon-remover-distance.hh,
* vcsn/algos/epsilon-remover-separate.hh,
* vcsn/algos/epsilon-remover.hh, vcsn/algos/eval.hh,
* vcsn/algos/grail.hh, vcsn/algos/has-bounded-lag.hh,
* vcsn/algos/is-eps-acyclic.hh, vcsn/algos/is-synchronized.hh,
* vcsn/algos/letterize.hh, vcsn/algos/lightest.hh,
* vcsn/algos/minimize-moore.hh, vcsn/algos/minimize-signature.hh,
* vcsn/algos/minimize-weighted.hh, vcsn/algos/shortest.hh,
* vcsn/algos/sort.hh, vcsn/algos/synchronize.hh,
* vcsn/algos/to-expression.hh:
Keep shared-ptrs, not const-refs to them.
Fix a few minor stylistic issues.