ps: ignore processes we are not allowed to look at
On OS X Yosemite, the cmdline of some processes cannot be read.  I
don't understand this.

For instance:

  psutil.Process(pid=87141, name='dbfseventsd')
  Fail:  psutil.Process(pid=87141, name='dbfseventsd') akim

yet, using `ps`, this process is reported to below to root, not to
akim as displayed here.  So we should have skipped it.

Time permitting, there might be a bug report to write here.

* libexec/vcsn-ps: here.