copy: improve copy_path
* vcsn/algos/copy.hh (copier): It's really a waste of resources to
iterate on all the transitions and on all states to copy just
transitions.  Just imagine the case of a large automaton, but just a
small part of which we are interested in.

These changes show that more is needed.  In particular, we should
really base everything on ranges rather than containers.  That's how
predicates should be handled: bounce to the range case using the
predicate to filter it.

Since we want that both for states and transitions, it will be easier
once transition_t and state_t are different.