trie: accept an argument stating the format of the file
Currently if we build a trie from a file, them the lines of the file
are read as if each line were a monomial.  So for instance `<` is
considered introducing a weight.

We need an argument stating the syntax of the stream.  Currently, in
addition to `"monomials"` (which was the previous implicit value), we
add `"words"` which means that each line must be understood as a raw
string up (and excluding) to end-of-line.

This allows to read safely /usr/share/dict/words for instance.

* vcsn/algos/trie.hh (trie_builder::add_words, trie_builder::add_monomials):
New.
(trie, cotrie): Now accept a format argument.
* vcsn/dyn/algos.hh: Likewise.
* python/vcsn_cxx.cc: Adjust.