polynomials: allow letter classes with mixed weights
Currently, by design, `a+<2>[^a]` is written without classes: only
`\e` is allowed to have a different weight for classes to be used.  In
some cases, such as the previous example, this is a nuisance,
especially on real world alphabet sizes.

A real world example is the edit-automaton, a tropical automaton,
which is, roughly, generated from `<0>a+<1>[^a]`: keeping `a` costs
nothing, but any other letter costs 1.

There is no visible performance regression on dot serialization,
although I expected one.

* vcsn/weightset/polynomialset.hh (print_without_classes_): Cluster the
labels per weight.

* tests/python/automaton.dir/lan_char_z.daut,
* tests/python/automaton.dir/lan_char_z.out.gv,
* tests/python/automaton.dir/lan_char_z.tex,
* tests/python/automaton.dir/lan_char_z.tex.gv: Adjust.