minimize: string algo redirects to tag dispatch
Use actual type to avoid static_assert.
`determinized_automaton<codeterminized_automaton<decltype(transpose(a))>>`
Use decltype on transpose because transpose(transpose(a)) returns a.
Hence, if a is a transpose the types would be different.
Move dynamic error handling overload to dyn so that an invalid call
to a minimize algorithm will raise in dyn and won't compile in static.

* vcsn/algos/minimize-hopcroft.hh: Here.
* vcsn/algos/minimize-moore.hh: Here.
* vcsn/algos/minimize-signature.hh: Here.
* vcsn/algos/minimize-brzozowski.hh: Here.
* vcsn/algos/minimize.hh: Here.