traits: make them SNIFAE compliant
Currently when a function template uses some of these traits, e.g.,
`weight_t_of<Aut>`, then it's a hard error when `Aut` is `int` for
instance.  This prevents SNIFAE from eliminating this obviously
invalid case.  We need a SNIFAE transparent implementation of
`weight_t_of<Aut>`.

This is needed by the following commit, otherwise we have a conflict
between the two left_mult: the one for automata, the other one for
valuesets (e.g., for expressions).  In that case, when instantiate for
Z, the compiler first tries the overload meant for automata, and
invokes `weight_t_of<int>`, which fails with a hard error, instead of
just discarding that overload.

* vcsn/ctx/traits.hh: Make it happen.