expression: provide a means to name expressions
This commit is actually pretty old, but was kept aside because, after
discussing with people such as Alexandre Duret-Lutz, it seemed nicer
to be able to _recognize_ an expression who has been given a name,
even if it's not the same instance.  Indeed, the approach of this
commit is to add `name` node in the AST of the expressions, and
therefore `(?<foo>a+b)` and `a+b` will not be recognized as the same
expression named foo.

However, it is very handy to be able to name an expression, especially
when using derived_term (or expansion) with non-trivial expressions.
And in that context, exact equality does have its own value as opposed
to "accidental" equality with some other subexpression which might be
equal.

So I don't know if we are going to keep this feature.  Hence, no doc.
It ought to be tested though...

* vcsn/core/rat/fwd.hh, vcsn/core/rat/expression.hh,
* vcsn/core/rat/expression.hxx (name): New type of expression node.

* vcsn/core/rat/expressionset.hh, vcsn/core/rat/expressionset.hxx
(name): New function.

* vcsn/core/rat/visitor.hh, vcsn/core/rat/printer.hh,
* vcsn/core/rat/printer.hxx (name): Support it.

* vcsn/algos/to-expansion.hh: Handle the named expressions: recover
the names.

* vcsn/algos/name.hh, vcsn/dyn/algos.hh (dyn::name): New.
* lib/vcsn/rat/parse.yy: Use it for `(?<NAME>expression)`.
* lib/vcsn/rat/scan.ll (`(?<NAME>`): New pattern.

* python/vcsn_cxx.cc (expression.name): New.