doc: levenshtein: regen/fix
When regenerating this page, we get a different automaton for `edit`,
with 50 states, including `(0,0),2`, which does not appear in the
previous versions of this file.  It is not the first time I observed
this, so in the previous revisions of this file I deliberatedly did
not check-in the changes in the automaton, until I knew for sure what
the issue was.

It turns out that if we go back to the creation of this file, commit
d23ade42324079acd8c82288a9063a42444883d2, it is _already_ wrong!  And
the automaton does have 50 states, although the pictures has less.  So
I guess the notebook was partially changed before final commit and was
in an inconsistent state.

It would be really nice to be able to run the notebooks as a test
suite.

* doc/notebooks/context.levenshtein.ipynb: here.