score: lightest: fix it, and improve it
There are two issues fixed here.

One bug is that vcsn-score uses some kind of a cache to avoid
executing several times the set-ups that are identical.
Unfortunately, it does not work properly: it does not "see" changes in
the global variables.  Because of this, the first tests of `lightest`
were running on automata meant for `shortest` (the previous bench),
which takes for ages (the automaton is way too large for `lightest`).

The second issue is that we run `lightest` on Z-automata.  In some
case the preconditions are not checked (and actually it was shown that
in that case the result can be wrong, see issue #32), and in other
cases, because the preconditions are checked, the performance are
ruined.  For instance:

    0.71   N/A a.lightest(5)        # a = std([a-e]?{150}), c = [a-e] -> Z
     N/A  0.08 a.lightest(5)        # a = std([a-e]?{150}), c = [a-e] -> Nmin

It is exactly the same piece of code which is running.  The difference
is only due to:

    require(!has_lightening_cycle(aut_),
            "lightest(n > 1): requires automaton without lightening cycles");

which runs Bellman-Ford in the case of Z, and not in the case of Nmin.
So of course, let's not check the preconditions, but truly the
algorithm itself.

* libexec/vcsn-score: Use clearer `setup`s to avoid the cache-related
bug.
Prefer Nmin to Z for lightest-path computations.
* libexec/vcsn-score-compare: Adjust accordingly.
Use similar `number`s for `lightest` and `lightest_automaton`.