determinize: use only queues
Surprisingly, the boolean and weight determinizations used two
different queueing strategies: stack for the former, and queue for the
latter.  Benches show that queue is most adequate for both.

Boolean determinization (with improvements):

    s/s: stack / stack  (for the record, was never committed)
    s/q: stack / queue  (previous commit)
    q/q: queue / queue  (this commit)

     s/s   s/q   q/q
    0.33  0.37  0.31   # a = de_bruijn(13), 10x
    0.73  0.82  0.65   # a = de_bruijn(14), 10x
    0.67  0.73  0.57   # a = de_bruijn(16), 2x
    0.76  0.84  0.71   # a = ladybird(18)
    0.76  0.81  0.72   # a = ladybird(18), c = [a-zA-Z0-9] -> B
    6.95  6.98  6.31   # a = ladybird(21)
    1.29  1.38  1.20   # a = std(([^]+a){100}), c = [a-zA-Z0-9] -> B, 2x

Weighted determinization (where there is no improvement, of course):

     s/s   s/q   q/q
    1.08  1.13  1.07   # a = de_bruijn(13), 10x
    1.07  1.13  1.06   # a = de_bruijn(13), c = [abc] -> F2, 10x
    2.32  2.53  2.30   # a = de_bruijn(14), 10x
    2.30  2.51  2.32   # a = de_bruijn(14), c = [abc] -> F2, 10x
    2.14  2.30  2.12   # a = de_bruijn(16), 2x
    2.11  2.25  2.12   # a = de_bruijn(16), c = [abc] -> F2, 2x
    0.28  0.32  0.28   # a = std(([^]+a){100}), c = [a-zA-Z0-9] -> B, 2x

* vcsn/algos/determinize.hh (determinized_automaton_impl): Use a
queue, instead of a stack.

Of course this changes the order in which states are discovered, hence
the state numbers change.  It's quite nice that we use the same number
for both algorithms (that was not the case before).  I kinda regret
though that we keep a queue and not a stack, as I tend to use a stack
based approach when I teach :)

* tests/python/determinize.dir/de-bruijn-3-det.gv,
* tests/python/determinize.dir/de-bruijn-8-det.gv,
* tests/python/determinize.dir/ladybird-4-det.gv,
* tests/python/determinize.dir/ladybird-8-det.gv: Adjust expectations.