determinize: weighted: stop at final states, not at post
We still have two implementations of the determinization, and I think
this is wrong: we should have a single one, but different back ends to
implement the polynomials of states (linear combinations of states).

Both had different termination procedures: the boolean ends when we
reach a final state ("end of final"), while the weighted
determinization reaches post ("on on post").

The benches are mostly equivalent.  I repeated them several times,
because the differences seem to be most in the error margin...

          end on post          |         end on final         |
 1.05  1.07  1.05  1.05  1.04  | 1.06  1.06  1.06  1.06  1.06 | ("weighted") # a = de_bruijn(13), 10x
 1.06  1.07  1.06  1.06  1.06  | 1.04  1.05  1.04  1.05  1.05 | ("weighted") # a = de_bruijn(13), c = [abc] -> F2, 10x
 2.28  2.29  2.29  2.28  2.28  | 2.33  2.34  2.34  2.33  2.34 | ("weighted") # a = de_bruijn(14), 10x
 2.32  2.31  2.31  2.30  2.30  | 2.32  2.29  2.31  2.31  2.31 | ("weighted") # a = de_bruijn(14), c = [abc] -> F2, 10x
 2.11  2.12  2.11  2.10  2.09  | 2.14  2.13  2.14  2.17  2.14 | ("weighted") # a = de_bruijn(16), 2x
 2.13  2.12  2.13  2.13  2.11  | 2.13  2.12  2.12  2.12  2.12 | ("weighted") # a = de_bruijn(16), c = [abc] -> F2, 2x
 0.28  0.28  0.26  0.28  0.27  | 0.27  0.29  0.28  0.28  0.27 | ("weighted") # a = std(([^]+a){100}), c = [a-zA-Z0-9] -> B, 2x
 0.32  0.32  0.31  0.32  0.31  | 0.30  0.30  0.32  0.31  0.31 | ()      # a = de_bruijn(13), 10x
 0.34  0.32  0.32  0.32  0.31  | 0.33  0.33  0.33  0.33  0.33 | ()      # a = de_bruijn(13), c = [abc] -> F2, 10x
 0.67  0.68  0.67  0.66  0.66  | 0.65  0.64  0.66  0.65  0.65 | ()      # a = de_bruijn(14), 10x
 0.67  0.67  0.67  0.67  0.66  | 0.69  0.68  0.68  0.69  0.69 | ()      # a = de_bruijn(14), c = [abc] -> F2, 10x
 0.58  0.59  0.57  0.57  0.57  | 0.55  0.55  0.57  0.55  0.55 | ()      # a = de_bruijn(16), 2x
 0.59  0.57  0.57  0.57  0.57  | 0.58  0.59  0.58  0.71  0.59 | ()      # a = de_bruijn(16), c = [abc] -> F2, 2x
 0.74  0.74  0.73  0.76  0.73  | 0.71  0.71  0.72  0.70  0.70 | ()      # a = ladybird(18)
 0.74  0.74  0.74  0.74  0.73  | 0.70  0.70  0.71  0.70  0.71 | ()      # a = ladybird(18), c = [a-zA-Z0-9] -> B
 6.60  6.54  6.57  6.57  6.54  | 6.12  6.47  6.52  6.30  6.24 | ()      # a = ladybird(21)
 1.20  1.18  1.20  1.18  1.18  | 1.18  1.15  1.16  1.16  1.16 | ()      # a = std(([^]+a){100}), c = [a-zA-Z0-9] -> B, 2x

"End on final" has better figures on the Boolean determinization
(e.g., ladybird(21)).  "End on post" seems to have slightly better
results for weighted determinization (e.g., de_bruijn(14)), but maybe
not...

So let's keep a single approach, to facilitate the fusion of both
flavors, and keep the "end on final", even though "end on post" is
conceptually simpler.

One advantage of "end of final" is that in B, it suffices to see that
one transitions reaches post to set the state as final: we don't have
to compute the weight of the possible remaining transitions to post
(scalar_product does that, as it uses "intersects" from
Boost.DynamicBitset).

* vcsn/algos/determinize.hh (detweighted_automaton_impl): Implement "end on final".