compose: cache to-be-added transitions in a polynomial
In order to avoid having to call `add_transition` each time, we cache
the transitions we add using a polynomial. We conserve a polynomial
for each successor of src. In the we end create the transitions using
`new_transition` on each monomial of the conserved polynomial.

(1) Using new_transition
(2) Using add_transition
(3) Using polynomial and storing them in a map
(4) Using polynomial and storing them in an unordered_map
(5) Using polynomial and storing them in a flat_map

  (1)    (2)   (3)   (4)   (5)
0.23s  0.17s 0.16s 0.24s 0.23s : a.compose(a2)  # a2 = std(([a-z]|[a-z])?{20}), a = std(a|a), 10x
1.49s 52.32s 6.86s 8.67s 7.19s : a.compose(a)   # a = std(['(a,a)'-'(i,z)']{4})
0.32s  0.32s 0.38s 0.35s 0.35s : a.compose(a)   # a = thompson(['(a,a)'-'(i,z)']{4})

* vcsn/algos/compose.hh: Here.