dyn: automaton: change the dyn::automaton class
The goal is to use dyn::automaton directly as a class and not like a
shared_ptr anymore (implying removing abstraction as make_automaton).

As dyn::automaton is not a shared_ptr anymore it need to be include in
the fwd.hh or it will be see as incomplete class at compile time.

* vcsn/dyn/automaton.hh: Here.
* vcsn/dyn/fwd.hh: Change headers.
* python/vcsn_cxx.cc: Change constructor to set automaton to null_ptr by default.