conjunction: fix the repeated conjunction
Conjunction on lan requires that we insplit the automata (except the
first one).  However, insplit was called only in the dyn layer, so
vcsn::conjunction was producing incorrect results.  And since the
repeated conjunction called it, the repeated dyn::conjunction on lan
automata was wrong too.

* vcsn/algos/conjunction.hh: Move the implementation in detail::,
and use `using` to bring the public API in vcsn::.
(detail::conjunction): Be sure to always insplit.
(do_insplit): Remove, we don't need it: insplit already does the right
thing when there cannot be spontaneous transitions.

* tests/python/conjunction.py: Check that case.