dyn: project for expansions and expressions
Issue #47.

This work shows severe limitations with the focus_automaton on
products of lan.  We had to work around that by making a special case
in the test-suite for automata in tests/python/project.py.

This commit also shows that we lack a uniform treatment of tapes in
valuesets.  The current situation is not satisfying, in that there are
several different approaches.

* vcsn/core/rat/expressionset.hh (project): New.
* vcsn/core/rat/project.hh: New.
* lib/vcsn/algos/others.cc: Adjust.

* vcsn/algos/project.hh (project): New overloads for expressions and
expansions.
* vcsn/dyn/algos.hh: Ditto.
* python/vcsn_cxx.cc: Bind them.
* tests/python/project.py: Check them.
Reorder logically.

* vcsn/core/rat/copy.hh: Comment changes.
* vcsn/dyn/expansion.hh: Fix includes.