conjunction: don't use complete_ in strict mode
The `complete_` function only knows the state number, not its name,
which it has to recover via a lookup in `origins()`.  There is no
reason for this in strict computations.

No change in performances, however this change is needed if we stop
using bimaps in strict mode, and in this case, `origins()` can only be
called once the computation is completed.

* vcsn/algos/conjunction.hh (conjunction, ldiv, infiltration):
Use add_conjunction_transitions.
(complete_ldiv): Remove.