inductive: a new means to construct an automaton
This is an initial stab at #57: provide an automatonset.

Eventually we should be able to use automata as weights, but to this
end we need an automatonset first.  Yet, before using this
automatonset it would be nice to be able to check it.

Well, I have long desired a means to check derived_term against other
algorithms to build automata, except that it's the only algorithm so
far that supports all these operators.  So let's introduce a new
exp-to-aut algorithm that simply use the automatonset to build a given
automaton.

There remains to decide what the operations of the automatonset should
be: standard, general, deterministic, etc.?  Well, let it be
parameterized by a tag that specify the functions we want to use.

Let's start with standard_tag: at least we know that, on basic
rational expressions, e.standard() and e.inductive('standard') should
be isomorphic.

* vcsn/core/automatonset.hh: Fill the gaps.
* vcsn/algos/inductive.hh: New.
* vcsn/dyn/algos.hh (inductive): New.
* python/vcsn_cxx.cc: Adjust.
* tests/python/standard.py: Check that both `standard` are equal.