tests: make inductive('standard') the reference
We now have two implementations of "standard": the original one of
`standard()`, and the one of `inductive("standard")`.  The first one
is way faster but generates gaps in the state numbers, while the
second one does not, and supports more operators.

So use the second one for references from now on.  This will allow to
cover the extended operators.

For issue #57.

* tests/python/standard.py (check): Make `inductive()` the reference,
and `standard()` the alternative.
Check equivalence with derived-term.
Adjust the references.