labels: fix the support of label escapes in expressions
Currently we do not handle escapes properly in expressions.  For
instance, vcsn.B.expression('\x5c') fails because the labelset
receives the decoded '\' so it looks for the next character.

We should not try to decode these escape sequences, let's pass them to
the labelset.

* lib/vcsn/rat/scan.ll: No longer translate escapes, leave them
to the labelset.
* vcsn/misc/stream.hh (get_char): This routine is getting complex,
it makes little sense to keep it inline, so move its implementation to...
* lib/vcsn/misc/stream.cc: here.
Handle the usual escapes ('\n', '\x0f').

Then we have another problem: vcsn.B.expression('\x2b') is printed a
`+`, so we cannot read it back.  We need to improve the escaping of
special characters in expressions (but not in other cases, for
instance, I want to see `+` on transitions of an automaton, not `\+`).

So we need to escape character differently depending on the caller.
Let's handle that in misc::format.

* vcsn/misc/format.hh (meta, meta_): New.
* vcsn/core/rat/printer.hxx: Use it.

Now check that.

* tests/python/expression.py: Check that we do support round-tripping on
all the hexadecimal escapes from \x01 to \xfe.