expressions: complain properly when \xff is used
Currently we do raise an error message, but the nasty character is not
escaped, and apparently Boost.Python, or Python, don't like that and
don't deliver the error message.

So we need to escape it properly.  And that revealed another problem:
'\xff' is confused with -1 (int), which is displayed a
`<end-of-file>`.

* vcsn/alphabets/setalpha.hh (add_letter): Escape the culprit.

* vcsn/misc/escape.hh, lib/vcsn/misc/escape.cc: Fix the int vs char
mess.
Use a template to support the string-returning versions.

* tests/python/expression.py: Check that.