polynomials: add support for @
Based on work by Thibaud Michaud.

Currently polynomialset feature a compose function that works properly
only for polynomials of expressions.  We actually want a version that
works for plain words.

Given that it's not always possible, we should have a traits to detect
whether two labelsets or two contexts are composable.

* vcsn/ctx/traits.hh (are_labelsets_composable, are_composable): New.
* vcsn/weightset/polynomialset.hh (compose): Generalize.
* tests/python/polynomial.py: Check composition.