python: get rid of "series"
The concept of "series" was a first step in the generalization of the
different sets of identities applied to expressions.  Today, it is
superseded by the more regular API based on the identities.

* python/vcsn_cxx.cc (identities): Rename as...
(make_identities): this.
(expression::as_series, expression::is_series): Remove.
(expression::as_expression, expression::as_): Fuse into...
(expression::as): this.
* tests/python/expression.py, tests/python/transducer.py: Adjust.