python: generate the C++ OO bindings
Adding a new algorithms to dyn is really too much work, so we should
mechanize as much as we can.  Besides, the object-oriented dyn layer
might be interesting to some people, so maybe we should expose it.
And actually this layer could be exposed to tools such as SWIG to
target other languages.

This patch initiates the handling of both issues: it parses
dyn/algos.hh to generate the dynoo layer.

Eventually, we should also generate the Boost.Python bindings.

No performance change observed.

* build-aux/bin/bindings-gen: New.
* python/vcsn_cxx.cc: Remove the parts that are now generated.