dyn: translate: add cluster on accessibility
When we compile an algo just-in-time, it's not the algo that cost the
more but everything we need around it.
A way to optimize it is to make cluster: compile various algorithm with the
same environment.

A first improve is on one file of the testsuite, accessible.py:
  from: 321.01s user 9.97s system 99% cpu 5:32.49 total
  to  : 222.74s user 6.29s system 99% cpu 3:49.70 total

Need more cluster to see a real improve on the testsuite.

* lib/vcsn/dyn/translate.cc: Here.