compose: add support in expand
`expand` develops multiplication over addition, but also, as a new
feature that was added recently, conjunction over addition.  But it
has to be enhance each time we add a new operator, such as `|`, and
more recently, `@`.

But in reality, it is doing the same thing as `copy` does: rewrite an
expression with some changes, e.g., different weightset, or different
identities.  Except that currently "distributive" distributes only
multiplication, not conjunction.  If at some point we wish to also
distribute conjunction, then we should just add more identities to
expressionset, and possibly a finer grain to select these identities.

So let's get rid of the current implementation of expand, and make it
just a call to copy with different ids.

Unfortunately currently copy does not take into account the changes of
ids, probably for efficiency.  So we actually serialize and
deserialize the expression...  Eventually, we will fix copy to this
end (see issue #69).

* vcsn/core/rat/identities.hh, lib/vcsn/rat/identities.cc:
Add a ctor that accepts const char*, for convenience.
* vcsn/algos/expand.hh (expand): Reimplement as a simple rewrite.
* vcsn/core/rat/expressionset.hh (make_expressionset): New overload,
for convenience.
* python/vcsn/expression.py (identities, type): New.
* tests/python/expand.py: Adjust expectations to the fact that & is
no longer distributed.
Test other operators.