build: fix warnings
Clang's -Wdangling-else insists that we have braces to avoid a
pseudo-ambiguity.  That annoys me, and we could have use #pragmas to
avoid this issue, since the user might run the compiler with that
warning enabled.  But then we have to fight against pragma about
diagnostics that trigger warnings by compiler that don't know them...

I surrender.

* vcsn/algos/to-expansion.hh: Add braces to pacify clang.
* build-aux/bin/bindings-gen: Declare local functions `static`.