rat: transpose: factor the handling of variadic operations
There is needless code duplication, plus local modifications.  For
instance sum and product rely on there neutral element.  This is wrong
if, for some reason, the identities are disabled.

So let's factor, but the right way: without relying on the neutrals.

* vcsn/core/rat/transpose.hh (visit_): New.
(mul): Do not rely on the neutral.
* tests/python/transpose.py: Check that.

fixup! rat: transpose: factor the handling of variadic operations