multiply: introduce a specific type for exponents
There are two `multiply`: one for the binary multiplication, and the
other one for the repeated multiplication (aka power).  Unfortunately
when working with Z (or Zmin, etc.) the value is int and both
signature collide (without the compiler reporting it!).  Bad luck: the
compiler gives precedence to power, so today <3><3>a resolves in
<27>a.

Let's introduce a class to denote exponents to make the difference.
`exponent` is quite long, let's name it `to`, like in `x to the n`.
Fixes #68.

Suggested by Alexandre Duret-Lutz.

* vcsn/misc/to.hh: New.
* vcsn/algos/conjunction.hh, vcsn/algos/multiply.hh: Use it.
* tests/python/multiply.py: Check that case.
* vcsn/algos/tags.hh: Add SFINAE protection to please GCC 6:
`multiply(automaton, automaton)` in its decltype invokes this function,
which resulted in the hard error instead of SFINAE.