score: disable incredibly long bench of shortest
In 5e4e2784a25f2dd1b4c99c473ccc71390224953c, to address #79, the
implementation of shortest no longer uses the Yen k-shortest path
algorithm.  As a result, a bench-case now swallows so much memory that
it is killed by the kernel before it finishes.

Revert this commit once #99 is fixed.

* libexec/vcsn-score: Comment a case.