ambiguous_word: fix the implementation
The previous implementation was really dumb, and the test suite was
equally as dumb.  Improve both...

Fixes issue #104.

* vcsn/algos/is-ambiguous.hh (is_ambiguous_impl): New functor to fix
the implementation of...
(ambiguous_word): this.
Don't traverse the original automaton: we really want to traverse the
conjunction.  Now we do enter the loops that introduce ambiguity.
* tests/python/is-ambiguous.py: Use a better test case.
The previous implementation returned `abc` instead of `abcdef` (plus
intro and outro).