derived-term: accept derivation for letterized multitape
So far we supported only free multitape labelset, because `[ab]|[xy]`
is simply `a|x + a|y + b|x + b|y`.  To support letterized we need to
be able to list the generators of `[ab]? | [xy]?` which are
  `\e|x + \e|y + a|\e + b|\e + a|x + a|y + b|x + b|y`
which is exactly what the previous commit fixed.

* vcsn/algos/derived-term.hh: Relax the constraints, the previous
patch took care of the rest.

* doc/notebooks/expression.derived_term.ipynb: Update to show
multitape support.

* tests/python/derivation.py: Check multitape.
* tests/python/to-expansion.py: Likewise.