tupleset: fix the generation of generators
Currently, the generators of `{a}? | {x}?` are `{a|x}`.  Make it
`{\e|x, a|\e, a|x}`.

* vcsn/labelset/tupleset.hh (generators): On letterized valuesets
that have a one, issue generators that contain this one.
(genset_t): Revise its definition.
I am no longer able to specify with genset_t is, it's too complex.
Yet, I cannot declare it inside tupleset using `decltype`, because
the compiler fails to resolve name for classes that are still no
completely known.  So the complexity is pushed on the caller side :(
* vcsn/core/rat/expressionset.hxx: For some reason, I can no longer
use boost::find: it ends in a nasty overload ambiguity that I was
unable to reproduce on a small example to send a bug report.
So let's use std::find.
* tests/unit/label.cc (check_generators): New.