style: remove random-label.hh
Because random-label covers all the labelsets, it includes them all.
As a consequence, instantiate.hh also contains them all, so touching
`tupleset.hh` triggers a recompilation of `lal, b`, which is painful
and totally useless.

Fixes #101.  This is not enough though: tupleset.hh is still included
by other algos.  The monolithic `instantiate.hh` is really a problem.

* vcsn/algos/random-label.hh:
Remove, dispatch its contents to...
* tests/bin/ipynbdoctest.py, vcsn/algos/random-automaton.hh,
* vcsn/algos/random-expression.hh, vcsn/core/rat/expressionset.hh,
* vcsn/labelset/labelset.hh, vcsn/labelset/nullableset.hh,
* vcsn/labelset/oneset.hh, vcsn/labelset/tupleset.hh,
* vcsn/labelset/wordset.hh:
here.