style: lightest: clean up
There are several vectors of transitions_t that are used, and it is
very unclear to what exactly they correspond.  At least we should have
typedefs, but I believe that it would be even better to have dedicated
types (instead of merely type aliases) to avoid type-compatibility for
things of a different nature.o

* vcsn/algos/k-lightest-path.hh (predecessors_t_of): New.
Use it.
Formatting changes.
* vcsn/algos/lightest-path.hh: Typo.
* vcsn/algos/lightest.hh: Simplify.