doc: ipynbdoctest: prepare for use with Automake
Issue #42.

* tests/bin/ipnbdoctest.py: Rename as...
* tests/bin/ipynbdoctest.py: this, because really, I get it
wrong too often.  Let's stick to the usual extension: ipynb.
Fix comment conventions.
Run the tests where the notebook is.
(--tap): New option.
Display TAP messages.
(canonicalize): Disable the normalization of the paths, as
it is so greedy that entire SVG renderings are smashed into a
single (short) line: we were completely failing to notice incorrect
automata (with different number of states for instance).

* doc/notebooks/local.mk: Set up a new test-suite category to check
the notebooks.
(check-nb): New shortcut.