has-bounded-lag: adjust score
The commit f893b2f had a huge impact on has-bounded-lag:

     0     1
    0.67  0.00 a.has_bounded_lag()  # a = std(['(a,x)'-'(b,y)']*{600}), 20x
    1.18  0.24 a.has_bounded_lag()  # a = std(['(a,x)'-'(b,y)']{1000}*), 10x
     0     1
     0. v2.3-041-g869ee51 style: minor improvements
     1. v2.3-042-gf893b2f tupleset: fix the generation of generators

But it's only because it changed unexpectedly the automaton on which
has-bounded-lag was checked: `['(a,x)'-'(b,y)']` changed from denoting
`a|x+a|y+a|z+b|x+b|y` to denoting `a|x+a|y+a|z+b|\e+b|x+b|y`, i.e., it
added `b|\e`.

See #119.

* libexec/vcsn-score: Use a safer expression.
Disable a bench on compose which suffers from the same condition.