doc: fix errors with dictionaries with quotes
See #108:

    lang = c.cotrie(filename='/usr/share/dict/words').determinize().strip()

    RuntimeError: expected ', got: aujourd'hui

because the French dictionary contains

    ...
    augustes
    aujourd'
    aujourd'hui
    aumône
    ...

The problem can be shrunk down to:

    vcsn.B.trie('''
    a'
    b'c
    ''')

But the latter is wrong: the default format is 'monomials', which
means that we do parse each line using our rules for mononials,
including the special handling of single quotes.  The format 'words'
(which could be named 'raw') is there for this use case.

So actually the real fix is that Spell-checker should use the right
format.

Fixes #108, and improves #42.

* doc/notebooks/Spell-checker.ipynb: Fix usage of context.cotrie.
Minor improvements.
* doc/notebooks/context.cotrie.ipynb, doc/notebooks/context.trie.ipynb:
Fix incorrect references.